+++ PAM/Controllers/NewRequestController.cs
@@ -20,16 +20,21 @@ namespace PAM.Controllers
         private readonly IADService _adService;
         private readonly UserService _userService;
         private readonly AppDbContext _dbContext;
+        private readonly TreeViewService _treeService;
+        private readonly OrganizationService _orgService;
 
         private IHttpContextAccessor _httpContextAccessor;
         private ISession _session => _httpContextAccessor.HttpContext.Session;
 
         public NewRequestController(IADService adService, UserService userService,
-            AppDbContext context, IHttpContextAccessor httpContextAccessor)
+            AppDbContext context, TreeViewService treeService,
+            OrganizationService orgService, IHttpContextAccessor httpContextAccessor)
         {
             _adService = adService;
             _dbContext = context;
             _userService = userService;
+            _treeService = treeService;
+            _orgService = orgService;
             _httpContextAccessor = httpContextAccessor;
         }
 
@@ -105,7 +110,44 @@ namespace PAM.Controllers
 
             // Routing from RequestType to SelectUnit
             //return RedirectToAction("RequestInfo");
-            return RedirectToAction("PickUnit", "SelectUnit");
+            return RedirectToAction("SelectUnit");
+        }
+
+        [HttpGet]
+        public IActionResult SelectUnit()
+        {
+            var myTree = _treeService.GenerateTree();
+            ViewData["MyTree"] = myTree;
+
+            return View();
+            //return View("../SelectUnit/SelectUnit");
+        }
+
+        [HttpPost]
+        public IActionResult SelectUnit(int selectedUnit)
+        {
+            TempData["selectedUnit"] = selectedUnit;
+
+            return RedirectToAction("SelectSystems");
+        }
+
+        [HttpGet]
+        public IActionResult SelectSystems()
+        {
+            var systemsList = _orgService.GetRelatedSystems((int)TempData["selectedUnit"]);
+
+            // Line below breaks
+            //TempData["SystemsList"] = systemsList;
+
+            return View(systemsList);
+        }
+
+        [HttpPost]
+        public IActionResult SystemSelected()
+        {
+            // Need to create a RequestedSystem object, save in session, get requestId and create entry at end
+
+            return RedirectToAction("RequestInfo");
         }
 
         [HttpGet]
