+++ PAM/Views/NewRequest/SelectSystems.cshtml
@@ -11,7 +11,7 @@
 
 <body onload="javascript: stepIndicate()">
   <partial name="../Shared/_ProgressBar.cshtml" />
-  <form asp-controller="SelectUnit" asp-action="SystemSelected" method="post">
+  <form asp-controller="NewRequest" asp-action="SystemSelected" method="post">
     <div class="alert alert-info">
       Step 5 - Default Systems Associated with Unit
     </div>
