+++ PAM/Views/NewRequest/SelectUnit.cshtml
@@ -6,7 +6,7 @@
 </head>
 <body onload="javascript: stepIndicate()">
   <partial name="../Shared/_ProgressBar.cshtml" />
-  <form asp-controller="SelectUnit" asp-action="PickUnit" method="post">
+  <form asp-controller="NewRequest" asp-action="SelectUnit" method="post">
     <div class="alert alert-info">
       Step 4 - Select The Type of Job
     </div>
